requirements.txt
torch==1.10.1
torchvision==0.11.2
torchaudio==0.10.1
openai-whisper
ollama
numpy==1.26.4
halo